Implemented Turn Timer
